+++ packages/yourloops/components/menus/user-menu.tsx
@@ -27,7 +27,7 @@
 
 import React, { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { useHistory } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 
 import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown'
 import CancelIcon from '@mui/icons-material/Cancel'
@@ -73,7 +73,7 @@ function UserMenu(): JSX.Element {
   const { t } = useTranslation('yourloops')
   const { user, logout } = useAuth()
   const { classes: { menu, typography } } = classes()
-  const history = useHistory()
+  const navigate = useNavigate()
   const theme = useTheme()
   const isMobile: boolean = useMediaQuery(theme.breakpoints.only('xs'))
   const [tooltipText, setTooltipText] = useState<string>('')
@@ -102,7 +102,7 @@ function UserMenu(): JSX.Element {
   }
 
   const onClickSettings = (): void => {
-    history.push('/preferences')
+    navigate('/preferences')
     closeMenu()
   }