+++ packages/yourloops/test/integration/pages/signup-stepper/caregiver-signup.spec.tsx
@@ -25,7 +25,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { act, screen } from '@testing-library/react'
+import { act, screen, waitFor } from '@testing-library/react'
 import {
   getAccessTokenWithPopupMock,
   loggedInUserEmail,
@@ -58,7 +58,10 @@ describe('Signup stepper as caregiver', () => {
   })
 
   it('should be able to create a caregiver account', async () => {
-    renderPage('/')
+    const router = renderPage('/')
+    await waitFor(() => {
+      expect(router.state.location.pathname).toEqual('/complete-signup')
+    })
     checkFooter({ needFooterLanguageSelector: true })
     checkStepper()