+++ packages/yourloops/test/integration/pages/trends/trends-view-for-caregiver.spec.tsx
@@ -25,7 +25,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { screen } from '@testing-library/react'
+import { screen, waitFor } from '@testing-library/react'
 import { mockAuth0Hook } from '../../mock/auth0.hook.mock'
 import { mockTeamAPI } from '../../mock/team.api.mock'
 import { minimalTrendViewData, mockDataAPI } from '../../mock/data.api.mock'
@@ -57,7 +57,10 @@ describe('Trends view for caregiver', () => {
 
   it('should render correct layout', async () => {
     mockDataAPI(minimalTrendViewData)
-    renderPage(`/patient/${unmonitoredPatientId}/trends`)
+    const router = renderPage(`/patient/${unmonitoredPatientId}/trends`)
+    await waitFor(() => {
+      expect(router.state.location.pathname).toEqual(`/patient/${unmonitoredPatientId}/trends`)
+    })
     expect(await screen.findByTestId('patient-nav-bar')).toBeVisible()
     checkPatientNavBarAsCaregiver()
     await checkCaregiverLayout(`${firstName} ${lastName}`)